Separate magic code into base file and implementation of magics.
Ran a first solid pass with Pyflakes to clean the code.
